In replies, show a small inset thumbnail of the person being replied to.
Using the UserpicsDownloader class to also store the mapping between
a user and her userpic. On startup, we also load this mapping
when we load the QSettings file.

The image for the inset is picked up from the cache only. No specific
network request is made. If the profile image of the replied-to chap
is not available in the cache, we will simply fall back to showing the
author's profile image.